Reorganize Code
The previous app layout was growing really messy and I was getting issues with the current user state not being updated before the heatmap tried to fetch the data. I just made a new barebones project directory which looks v. ugly but the currentUser and heatmap data actually update properly. Just a disclaimer that the reason it all broke was probably me lol.